refactored filesystem datasets to use huggingface api
currently, filesystem datasets had to tokenize the files twice in order to retrieve the total amount of token necessary for allocating enough space for the memmap.
instead, the contents from local files are read and a huggingface datasetdict object is created with datasets.Dataset.from_gen. after that, the tokenization
process is exactly the same as with huggingface.