bugfix custom_bn1d
the idea of custom_bn1d was to multiply each row of the input prompt with one scalar of the weight tensor and add a bias.
in the case of the input having less rows than the weight and bias tensor, new rows were added with embeddings from the rows calculated
previously.